+++ CHANGELOG.md
@@ -1,14 +1,29 @@
-## 0.0.1
+## [0.0.1]
 
 * First release, see usage for details.
 
-## 0.0.2
+## [0.0.2]
 
 * Added example.
 
-## 0.3.0
+## [0.3.0]
 
 * Modified comments.
 
-## 0.4.0
+## [0.4.0]
 * Fixed bug.
+
+## [0.5.0]
+* Fixed bug
+* Modified interfaces
+* 1.'headerWidget' => 'headerWidgetBuilder'
+* 2.'footerWidget' => 'footerWidgetBuilder'
+
+## [0.6.0]
+* Added loadMoreWidgetBuilder property.
+
+## [0.7.0]
+* Fixed loadMoreWidgetBuilder property.
+
+## [0.8.0]
+* Removed required property limit.